requirements.txt
websocket-client
pandas
python-dotenv
binance-connector
requests
